Cleanup: relocate invoice-related QR code functionality from receipt modules to invoice modules
