+++ apps/common/receipt.py
@@ -8,7 +8,6 @@ from pyppeteer.launcher import launch
 import barcode
 from barcode.writer import ImageWriter, SVGWriter
 import segno
-from segno import helpers
 
 from main import external_url
 from models import event_year
@@ -105,22 +104,6 @@ def make_qrfile(data, **kwargs):
     return qrfile
 
 
-def make_epc_qrfile(payment, **kwargs):
-    qrfile = io.BytesIO()
-    # TODO: this isn't currently used. Need to fetch IBAN from payment.recommended_destination
-    # and name from somewhere - maybe config rather than hard-coding.
-    qr = helpers.make_epc_qr(
-        name="FIXME FIXME FIXME",
-        iban="FIXME FIXME FIXME",
-        amount=payment.amount,
-        reference=payment.bankref,
-        encoding=1,
-    )
-    qr.save(qrfile, **kwargs)
-    qrfile.seek(0)
-    return qrfile
-
-
 def qrfile_to_svg(qrfile):
     return Markup(qrfile.getvalue().decode("utf-8"))
 
@@ -128,19 +111,6 @@ def qrfile_to_svg(qrfile):
 def format_inline_qr(data):
     qrfile = make_qrfile(
         data,
-        kind="svg",
-        svgclass=None,
-        omitsize=True,
-        xmldecl=False,
-        svgns=False,
-        nl=False,
-    )
-    return qrfile_to_svg(qrfile)
-
-
-def format_inline_epc_qr(payment):
-    qrfile = make_epc_qrfile(
-        payment,
         kind="svg",
         svgclass=None,
         omitsize=True,
