+++ apps/payments/invoice.py
@@ -1,4 +1,5 @@
 from decimal import Decimal
+from io import BytesIO
 import logging
 import shutil
 import os.path
@@ -14,11 +15,13 @@ from flask import (
     send_file,
 )
 from flask_login import login_required, current_user
+from markupsafe import Markup
+from segno import helpers
 from sqlalchemy.sql.functions import func
 from wtforms import TextAreaField, SubmitField
 
 from main import external_url, db
-from ..common.receipt import format_inline_epc_qr, render_pdf
+from ..common.receipt import render_pdf
 from models.product import Product, PriceTier
 from models.purchase import Purchase
 from ..common.forms import Form
@@ -26,6 +29,39 @@ from . import get_user_payment_or_abort
 from . import payments
 
 logger = logging.getLogger(__name__)
+
+
+def make_epc_qrfile(payment, **kwargs):
+    qrfile = BytesIO()
+    # TODO: this isn't currently used. Need to fetch IBAN from payment.recommended_destination
+    # and name from somewhere - maybe config rather than hard-coding.
+    qr = helpers.make_epc_qr(
+        name="FIXME FIXME FIXME",
+        iban="FIXME FIXME FIXME",
+        amount=payment.amount,
+        reference=payment.bankref,
+        encoding=1,
+    )
+    qr.save(qrfile, **kwargs)
+    qrfile.seek(0)
+    return qrfile
+
+
+def qrfile_to_svg(qrfile):
+    return Markup(qrfile.getvalue().decode("utf-8"))
+
+
+def format_inline_epc_qr(payment):
+    qrfile = make_epc_qrfile(
+        payment,
+        kind="svg",
+        svgclass=None,
+        omitsize=True,
+        xmldecl=False,
+        svgns=False,
+        nl=False,
+    )
+    return qrfile_to_svg(qrfile)
 
 
 class InvoiceForm(Form):
